+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/HttpDataSourceTestEnv.java
@@ -16,12 +16,14 @@
 
 package androidx.media3.test.utils;
 
+import static androidx.media3.datasource.DataSpec.HTTP_METHOD_POST;
 import static androidx.media3.test.utils.WebServerDispatcher.getRequestPath;
 
 import android.net.Uri;
 import androidx.media3.common.util.UnstableApi;
 import androidx.media3.common.util.Util;
 import androidx.media3.datasource.HttpDataSource;
+import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableListMultimap;
 import com.google.common.collect.ImmutableMap;
@@ -97,6 +99,24 @@ public class HttpDataSourceTestEnv extends ExternalResource {
           .setExtraResponseHeaders(EXTRA_HEADERS)
           .build();
 
+  public static final WebServerDispatcher.Resource POST_EMPTY_REQUEST_BODY =
+      new WebServerDispatcher.Resource.Builder()
+          .setPath("/post/empty-request")
+          .setHttpMethod(HTTP_METHOD_POST)
+          .setData(TestUtil.buildTestData(/* length= */ 20, seed++))
+          .setExtraResponseHeaders(EXTRA_HEADERS)
+          .build();
+
+  public static final WebServerDispatcher.Resource POST_WITH_REQUEST_BODY =
+      new WebServerDispatcher.Resource.Builder()
+          .setPath("/post/with-body")
+          .setHttpMethod(HTTP_METHOD_POST)
+          .setRequestHeaders(
+              ImmutableListMultimap.of(HttpHeaders.CONTENT_TYPE, "application/octet-stream"))
+          .setRequestBody(TestUtil.buildTestData(/* length= */ 10, seed++))
+          .setData(TestUtil.buildTestData(/* length= */ 20, seed++))
+          .setExtraResponseHeaders(EXTRA_HEADERS)
+          .build();
   public static final WebServerDispatcher.Resource REDIRECTS_TO_RANGE_SUPPORTED =
       RANGE_SUPPORTED.buildUpon().setPath("/redirects/to/range/supported").build();
 
@@ -112,6 +132,8 @@ public class HttpDataSourceTestEnv extends ExternalResource {
             "range not supported, length unknown", RANGE_NOT_SUPPORTED_LENGTH_UNKNOWN),
         createTestResource("gzip enabled", GZIP_ENABLED),
         createTestResource("gzip forced", GZIP_FORCED),
+        createTestResource("post empty request", POST_EMPTY_REQUEST_BODY),
+        createTestResource("post with body", POST_WITH_REQUEST_BODY),
         new DataSourceContractTest.TestResource.Builder()
             .setName("302 redirect")
             .setUri(
@@ -163,7 +185,9 @@ public class HttpDataSourceTestEnv extends ExternalResource {
                 RANGE_NOT_SUPPORTED,
                 RANGE_NOT_SUPPORTED_LENGTH_UNKNOWN,
                 GZIP_ENABLED,
-                GZIP_FORCED)));
+                GZIP_FORCED,
+                POST_EMPTY_REQUEST_BODY,
+                POST_WITH_REQUEST_BODY)));
 
     redirectionServer.start();
     redirectionServer.setDispatcher(
@@ -198,6 +222,10 @@ public class HttpDataSourceTestEnv extends ExternalResource {
         new DataSourceContractTest.TestResource.Builder()
             .setName(name)
             .setUri(Uri.parse(originServer.url(resource.getPath()).toString()))
+            .setHttpMethod(resource.getHttpMethod())
+            .setRequestHeaders(
+                Maps.transformValues(resource.getRequestHeaders().asMap(), Joiner.on(", ")::join))
+            .setRequestBody(resource.getRequestBody())
             .setResponseHeaders(Maps.transformValues(EXTRA_HEADERS.asMap(), v -> (List<String>) v))
             .setExpectedBytes(resource.getData());
     if (resource.resolvesToUnknownLength()) {